+++ src/Exceptions/UnsupportedOperationException.php
@@ -1,5 +1,5 @@
 <?php
-namespace wapmorgan\UnifiedArchive;
+namespace wapmorgan\UnifiedArchive\Exceptions;
 
 class UnsupportedOperationException extends \Exception
 {